rename: 서비스 함수 디렉토리 my 에서 my-info 로 수정 (#206)
